Changing the Collection of VUT
Making it a Vector item to give it more flavor. It has negative starting item score as a scaling mechanism for prefixes/suffixes. I know that you made it impossible for a player to actually find items with negative score, but if you do not want it to start that way I will try and find a way to do that. I already tried using sentimentality to fix it but it actually makes the item unusable so I am hoping this will suffice.